fix tests, move types
