+++ server/repository/src/db_diesel/stock_line.rs
@@ -253,9 +253,9 @@ mod test {
             key: StockLineSortField::ExpiryDate,
             desc: Some(false),
         };
-        // Make sure NULLS are last
+        // Make sure NULLS are first
         assert_eq!(
-            vec![from_row(line1()), from_row(line2()), from_row(line3())],
+            vec![from_row(line3()), from_row(line1()), from_row(line2())],
             repo.query(Pagination::new(), None, Some(sort)).unwrap()
         );
         // Desc by expiry date
@@ -263,9 +263,9 @@ mod test {
             key: StockLineSortField::ExpiryDate,
             desc: Some(true),
         };
-        // Make sure NULLS are first
+        // Make sure NULLS are last
         assert_eq!(
-            vec![from_row(line3()), from_row(line2()), from_row(line1())],
+            vec![from_row(line2()), from_row(line1()), from_row(line3())],
             repo.query(Pagination::new(), None, Some(sort)).unwrap()
         );
     }
